gui/requirements.txt
pyside6
pandas
numpy
scikit-learn
PyOpenGL
PyOpenGL_accelerate
tensorflow
